requirements.txt
tritonclient[http]
numpy
Pillow